[BACKLOG-6049] Unit tests and API documentation on the classes pentaho.type.Item (and pentaho.type.Item.Meta)
